Remove one layer in the MCP code
